django_app/requirements.txt
attrs==17.3.0
awsebcli==3.12.1
bleach==2.1.2
blessed==1.14.2
botocore==1.8.15
cement==2.8.2
colorama==0.3.7
decorator==4.1.2
Django==1.8
django-extensions==1.9.8
djangorestframework==3.7.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
html5lib==1.0.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
pandocfilters==1.4.2
parso==0.1.1
pathspec==0.5.0
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
pytest==3.3.1
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.9.1
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
widgetsnbextension==3.0.8
